requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-filter==21.1
django-getenv==1.3.2
djangorestframework==3.13.1
importlib-metadata==4.11.2
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
zipp==3.7.0
